Changed holding vehicles when riding vehicles
If a player is holding a vehicle and rides a vehicle, the player will now only use their non-dominant hand (set in the options menu) to hold the vehicle, leaving their dominant hand to control the vehicle. The vehicle that they are holding will be adjusted to make the balance look right. The same thing happens when riding a pig and holding a carrot on a stick. If the player has a carrot on a stick in both hands, the player will prefer to hold the carrot on the stick with their main hand while holding the vehicle in their non-dominant hand. When riding any other mob, or if the player isn't holding a carrot on a stick, the player will just hold the vehicle like normal.